VIA-294 Refactor: Extract eligibility from vaccine page to dedicated component
Precursor for other vaccine pages containing some EliD data in the
future. Simplifies logic in the shared Vaccine component to make it
clearer which pieces show for which vaccines. Ensures that the
eligibility component and fallback are coupled together to guard against
rendering an eligibility fallback when the eligibility component wasn't
supposed to show in the first place. Refactoring of tests to follow in
next commit.